+++ R/tags.R
@@ -1280,26 +1280,64 @@ flattenTags <- function(x) {
 # By not calling `as.tags(x)`, tagFunctions are not evaluated and other items
 # are not converted.
 flattenTagsRaw <- function(x) {
-  if (isTag(x) || isTagEnv(x)) {
+  relocateHtmlDeps <- function(z, type) {
+    zDeps <- htmlDependencies(z)
+    zDepsLen <- length(zDeps)
+    # Return early if there are no dependencies
+    if (zDepsLen == 0) return(z)
+
+    # Append the incoming html deps to z's children
+    # Perform position insert to not lose attrs on z/z$children
+    switch(type,
+      "tag" = {
+        children <- z[["children"]]
+        childrenLen <- length(children)
+        if (is.null(children)) {
+          z[["children"]] <- zDeps
+        } else {
+          z[["children"]][(childrenLen + 1):(childrenLen + zDepsLen)] <- zDeps
+        }
+      },
+      "tagList" = {
+        zLen <- length(z)
+        z[(zLen + 1):(zLen + zDepsLen)] <- zDeps
+      },
+      stop("unknown type: ", type)
+    )
+    # Remove html deps on z, as they are now in the children
+    htmlDependencies(z) <- NULL
+
+    z
+  }
+
+  if (isTagEnv(x)) {
+    # For tags, wrap them into a list (which will be unwrapped by caller)
+    list(x)
+  } else if (isTag(x)) {
+    # Append individual html deps as children elements.
+    # Attributes are eaisly lost when unlisted or collected.
+    # Instead, use the _newer_/stable approach of adding the html dep as a direct child
+    x <- relocateHtmlDeps(x, type = "tag")
     # For tags, wrap them into a list (which will be unwrapped by caller)
     list(x)
   } else if (isTagList(x)) {
-    if (length(x) == 0) {
-      # Empty lists are simply returned
-      x
-    } else {
-      # For items that are lists (but not tags), recurse
-      ret <- unlist(lapply(x, flattenTagsRaw), recursive = FALSE)
-      # Copy over attributes put on the original list (ex: html deps)
-      mostattributes(ret) <- attributes(x)
-      ret
-    }
+    # For items that are lists (but not tags), recurse
+    ret <- unlist(lapply(x, flattenTagsRaw), recursive = FALSE)
+    # Copy over attributes put on the original list (ex: html deps, class)
+    mostattributes(ret) <- attributes(x)
+    # Append individual html deps into the final list from the flattened tags
+    # It does not work out well to add attributes to `ret`, as the html deps are not found by findDependencies()
+    # Instead, use the _newer_/stable approach of adding the html dep as a direct child
+    ret <- relocateHtmlDeps(ret, type = "tagList")
+    # Return the list of items
+    ret
   } else {
     # This will preserve attributes if x is a character with attribute,
     # like what HTML() produces
     list(x)
   }
 }
+
 
 
 combineKeys <- function(x) {